backend/requirements.txt
flask
numpy
opencv-python
pillow
flask-cors
python-dateutil
torch
torchvision
